# swaggerDocumentation.yaml
openapi: 3.0.0
info:
  title: Ticketline API
  description: Documentation for the Ticketline API
  version: 0.0.1
servers:
  - url: localhost
    description: Can be used when spring application is started locally
paths:
...
  /users:
    get:
      tags:
        - userManagement
      summary: Gets a list of users
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/filterLocked"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      tags:
        - userManagement
      summary: Registers an user with the given data.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserWithPassword"
      responses:
        201:
          description: Successful creation of an user
          headers: 
            Location:
              $ref: "#/components/headers/Location"
        422:
          description: Validation failed for an input
        500:
          $ref: "#/components/responses/500-internal-error"
  /users/{id}:
    put:
      tags:
        - userManagement
      summary: Updates information of an user with the given data.
      parameters:
        - $ref: "#/components/parameters/userId"
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserWithPassword"
      responses:
        204:
          description: Successful update of an user.
        401:
          $ref: "#/components/responses/401-not-authenticated"
        422:
          description: Validation failed for an input
        500:
          $ref: "#/components/responses/500-internal-error"
    get:
      tags:
        - userManagement
      summary: Retreives information of an user with the given data.
      parameters:
        - $ref: "#/components/parameters/userId"
      security:
        - BearerAuth: []
      responses:
        200:
          description: Successful retreival of an user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        401:
          $ref: "#/components/responses/401-not-authenticated"
        404:
          description: The user with the given ID was not found
        500:
          $ref: "#/components/responses/500-internal-error"
    delete:
      tags:
        - userManagement
      summary: Deletes the user with the given ID.
      description: Only an administrator or the user itself can delete a given user account.
      parameters:
        - $ref: "#/components/parameters/userId"
      security:
        - BearerAuth: []
      responses:
        204:
          description: Successful deletion of an user.
        401:
          $ref: "#/components/responses/401-not-authenticated"
        403:
          $ref: "#/components/responses/403-forbidden"
        404:
          description: The user with the given ID was not found
        500:
          $ref: "#/components/responses/500-internal-error"
...
components:
  parameters:
    userId:
      name: id
      in: path
      description: ID of the user that is retreived
      required: true
      schema:
        type: integer
...
    filterLocked:
      name: filterLocked
      in: query
      description: Only return locked users if this is set to true
      required: false
      schema:
        type: boolean
...
  responses:
    500-internal-error:
      description: Internal Server Error
    401-not-authenticated:
      description: The user is not logged in
    403-forbidden:
      description: The user needs administrative rights
...
  headers:
    Location:
      required: true
      description: The location of the created resource
      schema:
        type: string
        format: uri
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    Gender:
      type: string
      enum:
        - male
        - female
        - other
...
    User:
      type: object
      properties:
        userId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        gender:
          $ref: "#/components/schemas/Gender"
        street:
          type: string
        zipCode:
          type: string
        city:
          type: string
        country:
          type: string
        isLocked:
          type: boolean
      required:
        - userId
        - firstName
        - lastName
        - email
        - gender
        - street
        - zipCode
        - city
        - country
        - isLocked
    UserWithPassword:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        gender:
          $ref: "#/components/schemas/Gender"
        street:
          type: string
        zipCode:
          type: string
        city:
          type: string
        country:
          type: string
        password:
          type: string
      required:
        - firstName
        - lastName
        - email
        - gender
        - street
        - zipCode
        - city
        - country
        - password
...